requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
cryptography==43.0.3
Django==5.2.5
dj-database-url==2.2.0
django-allauth==0.63.3
django-mathfilters==1.0.0
django-solo==2.5.1
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
django-widget-tweaks==1.5.0
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2
celery==5.4.0
redis==5.0.8
gunicorn==23.0.0
idna==3.10
packaging==25.0
pillow==11.3.0
PyJWT==2.8.0
reportlab==3.6.13
requests==2.32.4
sqlparse==0.5.3
django-jazzmin==2.6.0
typing_extensions==4.14.1
openai>=1.0.0
openpyxl>=3.1.0
urllib3==2.2.3
psycopg[binary]==3.2.1
uvicorn==0.30.6
xhtml2pdf==0.2.11
python-dotenv==1.0.1
whitenoise==6.7.0
fastapi==0.115.2
kafka-python==2.0.2
pandas==2.2.3
scikit-learn==1.5.2
pymupdf==1.24.10



